feat: first pass of large refactor to registering and supporting root namespaces based on a config value
