+++ common/src/main/java/dev/ftb/mods/ftbessentials/FTBEssentialsCommon.java
@@ -2,9 +2,6 @@ package dev.ftb.mods.ftbessentials;
 
 import dev.ftb.mods.ftbessentials.net.UpdateTabNameMessage;
 
-/**
- * @author LatvianModder
- */
 public class FTBEssentialsCommon {
 	public void updateTabName(UpdateTabNameMessage packet) {
 	}
