+++ common/src/main/java/dev/ftb/mods/ftbessentials/api/event/RTPEvent.java
@@ -4,9 +4,6 @@ import net.minecraft.core.BlockPos;
 import net.minecraft.server.level.ServerLevel;
 import net.minecraft.server.level.ServerPlayer;
 
-/**
- * @author LatvianModder
- */
 public class RTPEvent {
 	private final ServerLevel level;
 	private final ServerPlayer serverPlayer;
