+++ common/src/main/java/dev/ftb/mods/ftbessentials/commands/groups/AdminCommands.java
@@ -1,7 +1,11 @@
 package dev.ftb.mods.ftbessentials.commands.groups;
 
 import dev.ftb.mods.ftbessentials.commands.FTBCommand;
+import dev.ftb.mods.ftbessentials.commands.SimpleConfigurableCommand;
+import dev.ftb.mods.ftbessentials.commands.impl.admin.NicknameForCommand;
+import dev.ftb.mods.ftbessentials.config.FTBEConfig;
 import dev.ftb.mods.ftbessentials.util.OtherPlayerInventory;
+import net.minecraft.commands.arguments.EntityArgument;
 import net.minecraft.network.chat.Component;
 import net.minecraft.server.level.ServerPlayer;
 import net.minecraft.world.MenuProvider;
@@ -13,12 +17,26 @@ import net.minecraft.world.inventory.MenuType;
 
 import java.util.List;
 
+import static net.minecraft.commands.Commands.argument;
+import static net.minecraft.commands.Commands.literal;
+
 public class AdminCommands {
     public static final List<FTBCommand> COMMANDS = List.of(
+        // Manage nicknames
+        new NicknameForCommand(),
 
+        // Invsee command
+        new SimpleConfigurableCommand(
+                FTBEConfig.INVSEE,
+                literal("invsee")
+                        .requires(FTBEConfig.INVSEE.enabledAndOp())
+                        .then(argument("player", EntityArgument.player())
+                                .executes(context -> viewInventory(context.getSource().getPlayerOrException(), EntityArgument.getPlayer(context, "player")))
+                        )
+        )
     );
 
-    public static void viewInventory(ServerPlayer source, ServerPlayer player) {
+    public static int viewInventory(ServerPlayer source, ServerPlayer player) {
         source.openMenu(new MenuProvider() {
             @Override
             public Component getDisplayName() {
@@ -30,5 +48,7 @@ public class AdminCommands {
                 return new ChestMenu(MenuType.GENERIC_9x5, id, playerInventory, new OtherPlayerInventory(player), 5);
             }
         });
+
+        return 1;
     }
 }
