+++ common/src/main/java/dev/ftb/mods/ftbessentials/commands/groups/FeatureCommands.java
@@ -1,0 +1,16 @@
+package dev.ftb.mods.ftbessentials.commands.groups;
+
+import dev.ftb.mods.ftbessentials.commands.FTBCommand;
+import dev.ftb.mods.ftbessentials.commands.impl.KitCommand;
+
+import java.util.List;
+
+/**
+ * This class is for registering general feature commands where the commands are exclusively
+ * grouped with-in their own namespace. Kit commands are an example of this.
+ */
+public class FeatureCommands {
+    public static final List<FTBCommand> COMMANDS = List.of(
+        new KitCommand()
+    );
+}
