+++ common/src/main/java/dev/ftb/mods/ftbessentials/util/OtherPlayerInventory.java
@@ -5,9 +5,6 @@ import net.minecraft.world.Container;
 import net.minecraft.world.entity.player.Player;
 import net.minecraft.world.item.ItemStack;
 
-/**
- * @author LatvianModder
- */
 public class OtherPlayerInventory implements Container {
 	private final ServerPlayer player;
